refactor(node): sync node and seed implementation
